+++ @narative/gatsby-theme-novela/src/sections/article/Article.SEO.tsx
@@ -34,13 +34,11 @@ const ArticleSEO: React.FC<ArticleSEOProps> = ({
   imagelocation,
 }) => {
   const results = useStaticQuery(siteQuery);
-  const name = results.allSite.edges[0].node.siteMetadata.name;
   const siteUrl = results.allSite.edges[0].node.siteMetadata.siteUrl;
 
-  const authorsData = authors.map(author => ({
-    '@type': 'Person',
-    name: author.name,
-  }));
+  const authorsName = authors.map(author => (author.name));
+  const authorsSlug = authors.map(author => (author.slug));
+  const authorsBio = authors.map(author => (author.bio));
 
   // Checks if the source of the image is hosted on Contentful
   if (`${article.hero.seo.src}`.includes('ctfassets')) {
@@ -49,55 +47,21 @@ const ArticleSEO: React.FC<ArticleSEOProps> = ({
     imagelocation = `${siteUrl + article.hero.seo.src}`;
   }
 
-  /**
-   * For some reason `location.href` is undefined here when using `yarn build`.
-   * That is why I am using static query `allSite` to get needed fields: name & siteUrl.
-   */
-  let microdata = `{
-    "@context": "https://schema.org",
-    "@type": "Article",
-    "mainEntityOfPage": {
-      "@type": "WebPage",
-      "@id": "${siteUrl + location.pathname}"
-    },
-    "headline": "${article.title}",
-    "image": "${imagelocation}",
-    "datePublished": "${article.dateForSEO}",
-    "dateModified": "${article.dateForSEO}",
-    "author": ${JSON.stringify(authorsData)},
-    "description": "${article.excerpt.replace(/"/g, '\\"')}",
-    "publisher": {
-      "@type": "Organization",
-      "name": "${name}",
-      "logo": {
-        "@type": "ImageObject",
-        "url": "${siteUrl}/icons/icon-512x512.png"
-      }
-    }
-  }
-`.replace(/"[^"]+"|(\s)/gm, function(matched, group1) {
-    if (!group1) {
-      return matched;
-    } else {
-      return '';
-    }
-  });
-  /**
-   * See here for the explanation of the regex above:
-   * https://stackoverflow.com/a/23667311
-   */
-
   return (
     <SEO
-      title={article.title}
+      authorName={authorsName}
+      authorsBio={authorsBio}
+      authorsSlug={authorsSlug}
+      canonicalUrl={article.canonical_url}
+      dateforSEO={article.dateForSEO}
       description={article.excerpt}
       image={imagelocation}
-      timeToRead={article.timeToRead}
+      isBlogPost={true}
+      articlepathName={siteUrl + location.pathname}
       published={article.date}
-      pathname={location.pathname}
-      canonicalUrl={article.canonical_url}
+      timeToRead={article.timeToRead}
+      title={article.title}
     >
-      <script type="application/ld+json">{microdata}</script>
     </SEO>
   );
 };