+++ www/CHANGELOG.md
@@ -3,6 +3,81 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.6](https://github.com/narative/gatsby-theme-novela/compare/www@1.5.5...www@1.5.6) (2020-05-12)
+
+**Note:** Version bump only for package www
+
+
+
+
+
+## [1.5.5](https://github.com/narative/gatsby-theme-novela/compare/www@1.5.4...www@1.5.5) (2020-05-09)
+
+**Note:** Version bump only for package www
+
+
+
+
+
+## [1.5.4](https://github.com/narative/gatsby-theme-novela/compare/www@1.5.3...www@1.5.4) (2020-05-08)
+
+**Note:** Version bump only for package www
+
+
+
+
+
+## [1.5.3](https://github.com/narative/gatsby-theme-novela/compare/www@1.5.2...www@1.5.3) (2020-05-07)
+
+**Note:** Version bump only for package www
+
+
+
+
+
+## [1.5.2](https://github.com/narative/gatsby-theme-novela/compare/www@1.5.1...www@1.5.2) (2020-05-07)
+
+**Note:** Version bump only for package www
+
+
+
+
+
+## [1.5.1](https://github.com/narative/gatsby-theme-novela/compare/www@1.5.0...www@1.5.1) (2020-05-06)
+
+**Note:** Version bump only for package www
+
+
+
+
+
+# [1.5.0](https://github.com/narative/gatsby-theme-novela/compare/www@1.4.6...www@1.5.0) (2020-05-05)
+
+
+### Features
+
+* **blog:** add Notion in social links ([da82a09](https://github.com/narative/gatsby-theme-novela/commit/da82a0919ab23f7cdf8794febb10d9be491b7210))
+
+
+
+
+
+## [1.4.6](https://github.com/narative/gatsby-theme-novela/compare/www@1.4.5...www@1.4.6) (2020-05-03)
+
+**Note:** Version bump only for package www
+
+
+
+
+
+## [1.4.5](https://github.com/narative/gatsby-theme-novela/compare/www@1.4.4...www@1.4.5) (2020-05-02)
+
+**Note:** Version bump only for package www
+
+
+
+
+
 ## [1.4.4](https://github.com/narative/gatsby-theme-novela/compare/www@1.4.3...www@1.4.4) (2020-04-30)
 
 **Note:** Version bump only for package www